requirements.txt
pillow==8.0.1
torch==1.8.0+cu111
torchvision==0.9.0+cu111
pandas==1.1.0
opencv-python==4.4.0.46
albumentations==0.5.1
pytorch-lightning==1.0.6
omegaconf==2.0.5
wandb==0.10.22
torchmetrics==0.2.0
imbalanced-learn==0.8.0
